bas-db: IMapperTemplate.insert() returns long.

+++ bas/layer-5/bas-db/src/main/java/net/bodz/bas/db/batis/IMapperTemplate.java
@@ -16,7 +16,7 @@ public interface IMapperTemplate<T, C>
 
     T selectNext(long id);
 
-    int insert(T obj);
+    long insert(T obj);
 
     void update(T obj);
